Changes from last time
add changelog notes

Try using lift generic

This works but I'd have to also make newtypes for all fields
and create an alternate representation.
I'll just use pprint for now and ask matt about it if it all works.

This makes it compile and the tests pass

Try to do typeable approach

this has issues however, it'd only work on types with a typable
implementation, but for business adt's this is rarely implemented.